chore(queue): upgrade to v0.3.0 and update codebase (#829)
* refactor: refactor message handling and update dependencies

- Update dependencies in `go.mod` to newer versions
- Add `github.com/jpillora/backoff` as an indirect dependency
- Change `queue.NewConsumer` to `queue.NewRing` in `main.go`
- Remove unnecessary type conversion for `WorkerNum` in `main.go`
- Add `Payload` method to `PushNotification` in `notification.go`
- Change `QueuedMessage` to `TaskMessage` in `SendNotification` function
- Modify JSON unmarshal to use `Payload` method in `SendNotification`
- Update `Run` function to use `TaskMessage` instead of `QueuedMessage`
- Update `server_test.go` to use `TaskMessage` instead of `QueuedMessage`
- Change data types in `status.go` for `BusyWorkers`, `SuccessTasks`, `FailureTasks`, and `SubmittedTasks`


* update


---------

+++ metric/metrics_test.go
@@ -18,6 +18,6 @@ func TestNewMetrics(t *testing.T) {
 	time.Sleep(10 * time.Millisecond)
 	defer q.Release()
 	m := NewMetrics(q)
-	assert.Equal(t, 2, m.q.SubmittedTasks())
-	assert.Equal(t, 2, m.q.SuccessTasks())
+	assert.Equal(t, uint64(2), m.q.SubmittedTasks())
+	assert.Equal(t, uint64(2), m.q.SuccessTasks())
 }
